web-app/requirements.txt
Django>=1.5
django-tastypie>=0.10.0
mimeparse==0.1.3
python-memcached==1.53
